% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{build_sct}
\alias{build_sct}
\title{Build an SCT from an ExpressionSet}
\usage{
build_sct(x, sample_field, condition_field)
}
\arguments{
\item{x}{An ExpressionSet (possibly reduced gene list)}

\item{sample_field}{A pdata field representing the unique samples (repeated entries).}

\item{condition_field}{A pdata field representing the condition to plot.}
}
\value{
A sct (sample-condition table).
}
\description{
This creates a sample-condition table from an ExpressionSet with pData.
}
\details{
An SCT is a set of gene expressions indicating specific samples and associated with
a specific condition. This structure allows plotting of individual gene expression by
sample and by condition.

sct stands for sample-condition table. Briefly, it collects information from an ExpressionSet
into a set of samples with specific conditions, together with gene expression. This is really
only useful for gene subset. Given the gene subset, you can then plot the genes by sample
and by condition.

This function takes an ExpressionSet, extracts two fields from the pData (sample_field
and condition_field) and builds the necessary structure for plotting.
}
