% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variance_by_row.R
\name{filter_variance_by_row}
\alias{filter_variance_by_row}
\title{Filter rows to those having high enough variance.}
\usage{
filter_variance_by_row(x, threshold = 0)
}
\arguments{
\item{x}{An ExpressionSet.}

\item{threshold}{Variance threshold to use.}
}
\value{
A matrix with filtered variance is returned.
}
\description{
The ExpressionSet is filtered such that only genes
with a variance larger than a specific value (given
by threshold) are kept.
}
