% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sample_plot.R
\name{generate_sample_sample_density_plot_list}
\alias{generate_sample_sample_density_plot_list}
\title{Generate a list of sample/sample density plots.}
\usage{
generate_sample_sample_density_plot_list(
  x,
  comparisons = NULL,
  main = "Sample-Sample comparisons"
)
}
\arguments{
\item{x}{A matrix of samples (columns)}

\item{comparisons}{A matrix of sample pairings (columns)}

\item{main}{A title for the plot(s)}
}
\value{
A list of plots.
}
\description{
This function is suitable for generating a bunch of sample-sample
plots that can then be ggarrange'd to plot out in bulk.
}
