% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_process_celfiles.R
\name{load_and_process_celfiles}
\alias{load_and_process_celfiles}
\title{Load and process a celfile dataset.}
\usage{
load_and_process_celfiles(data)
}
\arguments{
\item{data}{A tibble of samples}
}
\value{
An list of (celfiles, rma_exprs).
}
\description{
It requires a
table/tibble of samples with at least one column called
"filenames". These filenames will be read in as CEL files.
Any associated data in the table will be added as pData to
the AffyBatch.
}
\details{
This does quite a bit (by calling other functions). First, it
loads the data into an AffyBatch and includes metadata in the
object (associated with the input parameter). Then, RMA is used
on the sample set. RSI is calculated. And the resulting AffyBatch and
ExpressionSet are returned as an object from this function.
}
