% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_process_celfiles.R
\name{load_celfiles}
\alias{load_celfiles}
\title{Load an experimental CEL file dataset.}
\usage{
load_celfiles(data)
}
\arguments{
\item{data}{A tibble of filenames (and perhaps other metadata).}
}
\value{
An AffyBatch object.
}
\description{
This is custom for the RSI_CLIA project. It requires a
table/tibble of samples with at least one column called
"filenames". These filenames will be read in as CEL files,
any associated data in the table will be added as pData to
the AffyBatch.
}
