% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_plot.R
\name{mds_plot}
\alias{mds_plot}
\title{Createa a ggplot-based MDS plot of gene expression data.}
\usage{
mds_plot(x, main = "MDS Plot of samples")
}
\arguments{
\item{x}{An ExpressionSet to plot.}

\item{main}{A title for the plot.}
}
\value{
A ggplot representing an MDS plot.
}
\description{
Createa a ggplot-based MDS plot of gene expression data.
}
