% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sample_plot.R
\name{pair_with}
\alias{pair_with}
\title{Pair a sample (n) with all the other samples in the matrix m.}
\usage{
pair_with(n, m)
}
\arguments{
\item{n}{The index of sample to pair everything against.}

\item{m}{A matrix of samples (one column is one sample).}
}
\value{
A matrix of columns representing pairings to consider.
}
\description{
This function takes an integer n and a matrix m. The goal is to
generate all pairings of n with members of l (except n itself).
This is done in index space (e.g., if n=1, then it would provide
(1,2), (1,3), (1,4) for a 4-column matrix).
}
\details{
I suspect there is a function in R to do this, but I couldn't find
it quickly.
}
