% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{Create a PCA ggplot of gene expression data.}
\usage{
pca_plot(x, main = "PCA Plot of samples")
}
\arguments{
\item{x}{An ExpressionSet to plot.}

\item{main}{A title for the graph.}
}
\value{
A ggplot representing a PCA plot.
}
\description{
Create a PCA ggplot of gene expression data.
}
