% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{pivot_sct}
\alias{pivot_sct}
\title{Internal function to pivot a SCT}
\usage{
pivot_sct(sct, sample_field, condition_field)
}
\arguments{
\item{sct}{The un-pivoted sample-condition table.}

\item{sample_field}{Name of the sample field}

\item{condition_field}{Name of the condition field}
}
\value{
A pivoted sct suitable for graphing.
}
\description{
Internal function to pivot a SCT
}
