% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{plot_sct}
\alias{plot_sct}
\title{Plot a sample-condition table}
\usage{
plot_sct(sct, sample_field = NULL, condition_field = NULL, main = NULL)
}
\arguments{
\item{sct}{A sample-condition table.}

\item{sample_field}{The string name of the sample field}

\item{condition_field}{The string name of the condition field}

\item{main}{a title for the plot}
}
\value{
A ggplot showing the gene expression for a set of samples and conditions.
}
\description{
Plot the gene expression data within a sample-condition table, indicating both
the samples and conditions in the plot.
}
\details{
sct stands for sample-condition table. Briefly, it collects information from an ExpressionSet
into a set of samples with specific conditions, together with gene expression. This is really
only useful for gene subset. Given the gene subset, you can then plot the genes by sample
and by condition.

The samples are indicated by color and the conditions are indicates by shape. NOTE:
if the condition is not specified, then just the samples are plotted (i.e., no conditions).

Also note that the samples and conditions have to be factors. This will be done by default
for you. However, if you want a particular ordering of the levels then convert the column
to a factor before calling the function.

NOTE: If you want a list, use plot_sct_by_sample()
}
