% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{plot_sct_by_sample}
\alias{plot_sct_by_sample}
\title{Generate sct plots by sample as a list.}
\usage{
plot_sct_by_sample(
  sct,
  sample_field = NULL,
  condition_field = NULL,
  main = NULL
)
}
\arguments{
\item{sct}{The sample-condition table (tibble/data.frame)}

\item{sample_field}{The string representing the sample field}

\item{condition_field}{The string representing the condition field}

\item{main}{A title to use for plots (one is autogenerated).}
}
\value{
A list of ggplots.
}
\description{
Given a sample-condition table, generate a list of plots by sample. That is,
every sample generates a ggplot showing the gene expression of a subset of
genes for a specific condition (as highlighted by the sct, sample_field and
condition_field).
}
\details{
sct stands for sample-condition table. Briefly, it collects information from an ExpressionSet
into a set of samples with specific conditions, together with gene expression. This is really
only useful for gene subset. Given the gene subset, you can then plot the genes by sample
and by condition.
}
