% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_process_celfiles.R
\name{process_celfiles}
\alias{process_celfiles}
\title{Process CEL files by RMA normalization, then compute RSI.}
\usage{
process_celfiles(celfiles)
}
\arguments{
\item{celfiles}{An AffyBatch of CEL files.}
}
\value{
A list of celfiles, rma_exprs corresponding to the input. The pData of both include a new variable, called rsi.
}
\description{
Process CEL files by RMA normalization, then compute RSI.
}
