% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sample_plot.R
\name{sample_sample_density_plot}
\alias{sample_sample_density_plot}
\title{Generate a nice sample-sample density plot of two samples in matrix x.}
\usage{
sample_sample_density_plot(x, fc = 2, main = NULL)
}
\arguments{
\item{x}{Matrix of twp columns (samples) to graph.}

\item{fc}{Fold-change cutoff to consider (2.0 is default).}

\item{main}{A main title for plot}
}
\value{
Return a ggplot sample-sample density plot.
}
\description{
Generate a nice sample-sample density plot of two samples in matrix x.
}
