% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sample_plot.R
\name{sample_sample_plot}
\alias{sample_sample_plot}
\title{Create a sample-sample scatter plot.}
\usage{
sample_sample_plot(
  x,
  y,
  fc = 2,
  main = "Sample-Sample comparisons",
  xlab = "Sample 1",
  ylab = "Sample 2"
)
}
\arguments{
\item{x}{Sample 1}

\item{y}{Sample 2}

\item{fc}{Fold change lines}

\item{main}{Main title}

\item{xlab}{X label}

\item{ylab}{Y label}
}
\value{
ggplot of scatter plot.
}
\description{
This is a plot type that does a scatter plot of two samples, that
are expected to be highly similar. Fold-change lines are included,
as are basic statistics on the degree of differences between the
two samples and degree of similarity (correlation).
}
\details{
The assumption is that the plots are fairly similar, so there
should be a place to include text for metrics in the lower
right quadrant of the graph.
}
